restructure modules simulation pkgs
