+++ x/auth/simulation/operations/fake.go
@@ -1,4 +1,4 @@
-package simulation
+package operations
 
 import (
 	"errors"
