+++ x/distribution/simulation/decoder.go
@@ -1,4 +1,4 @@
-package decoder
+package simulation
 
 import (
 	"bytes"
