+++ x/mint/module.go
@@ -14,7 +14,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/x/mint/client/cli"
 	"github.com/cosmos/cosmos-sdk/x/mint/client/rest"
-	"github.com/cosmos/cosmos-sdk/x/mint/simulation/decoder"
+	"github.com/cosmos/cosmos-sdk/x/mint/simulation"
 )
 
 var (
@@ -72,7 +72,7 @@ type AppModuleSimulation struct{}
 
 // RegisterStoreDecoder registers a decoder for mint module's types.
 func (AppModuleSimulation) RegisterStoreDecoder(sdr sdk.StoreDecoderRegistry) {
-	sdr[StoreKey] = decoder.DecodeStore
+	sdr[StoreKey] = simulation.DecodeStore
 }
 
 //____________________________________________________________________________
