+++ x/slashing/module.go
@@ -15,7 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/slashing/client/cli"
 	"github.com/cosmos/cosmos-sdk/x/slashing/client/rest"
 	"github.com/cosmos/cosmos-sdk/x/slashing/internal/types"
-	"github.com/cosmos/cosmos-sdk/x/slashing/simulation/decoder"
+	"github.com/cosmos/cosmos-sdk/x/slashing/simulation"
 )
 
 var (
@@ -77,7 +77,7 @@ type AppModuleSimulation struct{}
 
 // RegisterStoreDecoder registers a decoder for slashing module's types
 func (AppModuleSimulation) RegisterStoreDecoder(sdr sdk.StoreDecoderRegistry) {
-	sdr[StoreKey] = decoder.DecodeStore
+	sdr[StoreKey] = simulation.DecodeStore
 }
 
 //____________________________________________________________________________
