+++ x/slashing/simulation/operations/msgs.go
@@ -1,4 +1,4 @@
-package simulation
+package operations
 
 import (
 	"fmt"
