+++ x/staking/simulation/decoder_test.go
@@ -1,4 +1,4 @@
-package decoder
+package simulation
 
 import (
 	"fmt"
