+++ x/supply/simulation/decoder_test.go
@@ -1,4 +1,4 @@
-package decoder
+package simulation
 
 import (
 	"fmt"
@@ -10,28 +10,32 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/mint/internal/types"
+	"github.com/cosmos/cosmos-sdk/x/supply/internal/keeper"
+	"github.com/cosmos/cosmos-sdk/x/supply/internal/types"
 )
 
 func makeTestCodec() (cdc *codec.Codec) {
 	cdc = codec.New()
 	sdk.RegisterCodec(cdc)
+	codec.RegisterCrypto(cdc)
+	types.RegisterCodec(cdc)
 	return
 }
-
 func TestDecodeStore(t *testing.T) {
 	cdc := makeTestCodec()
-	minter := types.NewMinter(sdk.OneDec(), sdk.NewDec(15))
+
+	totalSupply := types.NewSupply(sdk.NewCoins(sdk.NewInt64Coin(sdk.DefaultBondDenom, 1000)))
 
 	kvPairs := cmn.KVPairs{
-		cmn.KVPair{Key: types.MinterKey, Value: cdc.MustMarshalBinaryLengthPrefixed(minter)},
+		cmn.KVPair{Key: keeper.SupplyKey, Value: cdc.MustMarshalBinaryLengthPrefixed(totalSupply)},
 		cmn.KVPair{Key: []byte{0x99}, Value: []byte{0x99}},
 	}
+
 	tests := []struct {
 		name        string
 		expectedLog string
 	}{
-		{"Minter", fmt.Sprintf("%v\n%v", minter, minter)},
+		{"Supply", fmt.Sprintf("%v\n%v", totalSupply, totalSupply)},
 		{"other", ""},
 	}
